requirements.txt
bs4
PyYaml
qrcodes
pandas
pillow
